# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/documents": {
      "get": {
        "operationId": "DocumentsController_list",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 100,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "POLICY",
                "SOP",
                "WORK_INSTRUCTION",
                "FORM",
                "SPECIFICATION",
                "OTHER"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "DRAFT",
                "IN_REVIEW",
                "APPROVED",
                "EFFECTIVE",
                "OBSOLETE"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentSummaryResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Documents"
        ]
      },
      "post": {
        "operationId": "DocumentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocumentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetailResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Documents"
        ]
      }
    },
    "/api/v1/documents/{documentId}": {
      "get": {
        "operationId": "DocumentsController_get",
        "parameters": [
          {
            "name": "documentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetailResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Documents"
        ]
      }
    },
    "/api/v1/documents/{documentId}/versions": {
      "post": {
        "operationId": "DocumentsController_createVersion",
        "parameters": [
          {
            "name": "documentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocumentVersionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetailResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Documents"
        ]
      }
    },
    "/api/v1/documents/{documentId}/review-request": {
      "post": {
        "operationId": "DocumentsController_requestReview",
        "parameters": [
          {
            "name": "documentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestDocumentReviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetailResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Documents"
        ]
      }
    },
    "/api/v1/documents/{documentId}/review-decision": {
      "post": {
        "operationId": "DocumentsController_reviewDecision",
        "parameters": [
          {
            "name": "documentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentDecisionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetailResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Documents"
        ]
      }
    },
    "/api/v1/documents/{documentId}/approval-decision": {
      "post": {
        "operationId": "DocumentsController_approvalDecision",
        "parameters": [
          {
            "name": "documentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentDecisionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetailResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Documents"
        ]
      }
    },
    "/api/v1/users/invitations": {
      "post": {
        "operationId": "UserInvitationsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvitationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Invitations"
        ]
      },
      "get": {
        "operationId": "UserInvitationsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvitationResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Invitations"
        ]
      }
    },
    "/api/v1/users/invitations/{invitationId}/resend": {
      "post": {
        "operationId": "UserInvitationsController_resend",
        "parameters": [
          {
            "name": "invitationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Invitations"
        ]
      }
    },
    "/api/v1/users/invitations/{invitationId}": {
      "delete": {
        "operationId": "UserInvitationsController_revoke",
        "parameters": [
          {
            "name": "invitationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Invitations"
        ]
      }
    },
    "/api/v1/invitations/preview": {
      "post": {
        "operationId": "PublicInvitationsController_preview",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitationTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationPreviewDto"
                }
              }
            }
          }
        },
        "tags": [
          "Invitations"
        ]
      }
    },
    "/api/v1/invitations/accept": {
      "post": {
        "operationId": "PublicInvitationsController_accept",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInvitationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Invitations"
        ]
      }
    },
...
  },
  "info": {
    "title": "Qualyra GMP API",
    "description": "Secure multi-tenant quality management API.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "qualyra_refresh"
      }
    },
    "schemas": {
...
      "AuthenticationResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Short-lived RS256 access token."
          },
          "csrfToken": {
            "type": "string",
            "description": "Double-submit token for cookie-based operations."
          },
          "user": {
            "$ref": "#/components/schemas/AuthenticatedUserDto"
          },
          "tenant": {
            "$ref": "#/components/schemas/TenantSummaryDto"
          }
        },
        "required": [
          "accessToken",
          "csrfToken",
          "user",
          "tenant"
        ]
      },
...
      "Object": {
        "type": "object",
        "properties": {}
      },
      "DocumentUserSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          }
        },
        "required": [
          "id",
          "displayName",
          "email"
        ]
      },
      "DocumentVersionSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "versionNumber": {
            "type": "number",
            "minimum": 1
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "object",
            "nullable": true
          },
          "changeSummary": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "IN_REVIEW",
              "APPROVED",
              "EFFECTIVE",
              "SUPERSEDED",
              "OBSOLETE"
            ]
          },
          "createdBy": {
            "$ref": "#/components/schemas/DocumentUserSummaryDto"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "versionNumber",
          "title",
          "changeSummary",
          "status",
          "createdBy",
          "createdAt"
        ]
      },
      "DocumentSummaryResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "POLICY",
              "SOP",
              "WORK_INSTRUCTION",
              "FORM",
              "SPECIFICATION",
              "OTHER"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "IN_REVIEW",
              "APPROVED",
              "EFFECTIVE",
              "OBSOLETE"
            ]
          },
          "currentVersionNumber": {
            "type": "number",
            "minimum": 1
          },
          "owner": {
            "$ref": "#/components/schemas/DocumentUserSummaryDto"
          },
          "createdBy": {
            "$ref": "#/components/schemas/DocumentUserSummaryDto"
          },
          "currentVersion": {
            "$ref": "#/components/schemas/DocumentVersionSummaryDto"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "code",
          "type",
          "status",
          "currentVersionNumber",
          "owner",
          "createdBy",
          "currentVersion",
          "createdAt",
          "updatedAt"
        ]
      },
      "DocumentVersionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "versionNumber": {
            "type": "number",
            "minimum": 1
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "object",
            "nullable": true
          },
          "changeSummary": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "IN_REVIEW",
              "APPROVED",
              "EFFECTIVE",
              "SUPERSEDED",
              "OBSOLETE"
            ]
          },
          "createdBy": {
            "$ref": "#/components/schemas/DocumentUserSummaryDto"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "content": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "versionNumber",
          "title",
          "changeSummary",
          "status",
          "createdBy",
          "createdAt",
          "content"
        ]
      },
...
      "DocumentDetailResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "POLICY",
              "SOP",
              "WORK_INSTRUCTION",
              "FORM",
              "SPECIFICATION",
              "OTHER"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "IN_REVIEW",
              "APPROVED",
              "EFFECTIVE",
              "OBSOLETE"
            ]
          },
          "currentVersionNumber": {
            "type": "number",
            "minimum": 1
          },
          "owner": {
            "$ref": "#/components/schemas/DocumentUserSummaryDto"
          },
          "createdBy": {
            "$ref": "#/components/schemas/DocumentUserSummaryDto"
          },
          "currentVersion": {
            "$ref": "#/components/schemas/DocumentVersionSummaryDto"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "versions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocumentVersionResponseDto"
            }
          },
          "workflows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocumentWorkflowResponseDto"
            }
          }
        },
        "required": [
          "id",
          "code",
          "type",
          "status",
          "currentVersionNumber",
          "owner",
          "createdBy",
          "currentVersion",
          "createdAt",
          "updatedAt",
          "versions",
          "workflows"
        ]
      },
      "CreateDocumentDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "QMS-SOP-001"
          },
          "type": {
            "type": "string",
            "enum": [
              "POLICY",
              "SOP",
              "WORK_INSTRUCTION",
              "FORM",
              "SPECIFICATION",
              "OTHER"
            ],
            "example": "SOP"
          },
          "ownerUserId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "example": "Control of GMP documents"
          },
          "description": {
            "type": "string",
            "example": "Defines the controlled document lifecycle."
          },
          "content": {
            "type": "string",
            "example": "1. Purpose\nThis procedure defines..."
          },
          "changeSummary": {
            "type": "string",
            "example": "Initial controlled draft."
          }
        },
        "required": [
          "code",
          "type",
          "title",
          "content",
          "changeSummary"
        ]
      },
      "CreateDocumentVersionDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Control of GMP documents"
          },
          "description": {
            "type": "string",
            "example": "Defines the controlled document lifecycle."
          },
          "content": {
            "type": "string",
            "example": "1. Purpose\nThis revised procedure defines..."
          },
          "changeSummary": {
            "type": "string",
            "example": "Clarified document ownership."
          }
        },
        "required": [
          "title",
          "content",
          "changeSummary"
        ]
      },
      "RequestDocumentReviewDto": {
        "type": "object",
        "properties": {
          "reviewerUserId": {
            "type": "string",
            "format": "uuid"
          },
          "approverUserId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "reviewerUserId",
          "approverUserId"
        ]
      },
      "DocumentDecisionDto": {
        "type": "object",
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "APPROVE",
              "REJECT"
            ]
          },
          "comment": {
            "type": "string",
            "example": "Reviewed against the current GMP procedure."
          }
        },
        "required": [
          "decision",
          "comment"
        ]
      },
      "CreateInvitationDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "operator@acme.example"
          },
          "roleIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "email",
          "roleIds"
        ]
      },
...
      "InvitationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "ACCEPTED",
              "REVOKED",
              "EXPIRED"
            ]
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvitationRoleResponseDto"
            }
          },
          "invitedBy": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "acceptedAt": {
            "type": "object",
            "format": "date-time",
            "nullable": true
          },
          "lastSentAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "email",
          "status",
          "roles",
          "invitedBy",
          "expiresAt",
          "lastSentAt",
          "createdAt"
        ]
      },
      "InvitationTokenDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Opaque invitation token received by email."
          }
        },
        "required": [
          "token"
        ]
      },
      "InvitationPreviewDto": {
        "type": "object",
        "properties": {
          "tenantName": {
            "type": "string"
          },
          "tenantSlug": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvitationRoleResponseDto"
            }
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "tenantName",
          "tenantSlug",
          "email",
          "roles",
          "expiresAt"
        ]
      },
      "AcceptInvitationDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Opaque invitation token received by email."
          },
          "displayName": {
            "type": "string",
            "example": "Alex Morgan"
          },
          "password": {
            "type": "string",
            "minLength": 12,
            "maxLength": 128
          }
        },
        "required": [
          "token",
          "displayName",
          "password"
        ]
      },
...
    }
  }
}